Create google grants compliance criteria fail email - charity
Modular Compliance Checks: The script uses separate functions to check different aspects of compliance, making it easier to update and manage.

Immediate Alerting: The focus of this script is on alerting. As soon as any non-compliance is detected, it compiles the issues and sends an email alert.

Customizable Alert Content: The sendComplianceAlertEmail function can be customized to include more details or instructions based on the nature of the compliance issues.